Synchronize user->tv-show and user->tv-season list status
We want to both track the list status of each season and of the
tv show in general. This allows us to do a bunch of things:
- Calculate the average score for a given tv show (average of all
individual tv season ratings by the user for that tv show)
- Relevant for later: Display list status on the tv show details
page
- Especially important: now we can display new seasons of a tv show
that is on the users' list on the seasonal overview. Example: I've watched
the first 2 seasons of The Expanse. Obviously I would be interested in
The Expanse Season 3 showing up on my seasonal overview, even if I have
the `sequels` filter disabled. Previously we couldn't display the item
since technically `The Expanse Season 3` wasn't in the users' list.
Now we can, since we track list status for both tv seasons and tv shows,
synchronizing them automatically. In the future, we should also allow
users to manually update this status.